adapters: move the adapter base classes into lib/adapters from lib/runtime
This helps to consolidate the code related to implementing adapters into
lib/adapters. Note that the logic for loading / registering adapters still
resides in lib/runtime.
