+++ lib/adapters/adapter-read.js
@@ -2,7 +2,7 @@
 var _ = require('underscore');
 var EventEmitter = require('eventemitter3');
 var JuttleLogger = require('../logger');
-var juttle_utils = require('./juttle-utils');
+var juttle_utils = require('../runtime/juttle-utils');
 var errors = require('../errors');
 
 //
