+++ lib/adapters/adapter-write.js
