+++ lib/adapters/file/read.js
@@ -2,7 +2,7 @@
 
 var _ = require('underscore');
 var fs = require('fs');
-var AdapterRead = require('../../runtime/adapter-read');
+var AdapterRead = require('../adapter-read');
 var parsers = require('../parsers');
 var JuttleMoment = require('../../runtime/types/juttle-moment');
 var errors = require('../../errors');
