+++ lib/adapters/file/write.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var AdapterWrite = require('../../runtime/adapter-write');
+var AdapterWrite = require('../adapter-write');
 var Promise = require('bluebird');
 var fs = Promise.promisifyAll(require('fs'));
 var serializers = require('../serializers');
