+++ lib/adapters/http/read.js
@@ -2,7 +2,7 @@
 
 var _ = require('underscore');
 var contentType = require('content-type');
-var AdapterRead = require('../../runtime/adapter-read');
+var AdapterRead = require('../adapter-read');
 var parsers = require('../parsers');
 var errors = require('../../errors');
 var Promise = require('bluebird');
