+++ lib/adapters/http/write.js
@@ -2,7 +2,7 @@
 
 var _ = require('underscore');
 var request = require('request');
-var AdapterWrite = require('../../runtime/adapter-write');
+var AdapterWrite = require('../adapter-write');
 var errors = require('../../errors');
 
 // taken from https://gist.github.com/timruffles/3377784
