+++ lib/adapters/http_server.js
@@ -3,7 +3,7 @@
 var http = require('http');
 var _ = require('underscore');
 var contentType = require('content-type');
-var AdapterRead = require('../runtime/adapter-read');
+var AdapterRead = require('./adapter-read');
 var Promise = require('bluebird');
 var parsers = require('./parsers');
