+++ lib/adapters/stdio/read.js
@@ -1,7 +1,7 @@
 'use strict';
 /* eslint-env node */
 var _ = require('underscore');
-var AdapterRead = require('../../runtime/adapter-read');
+var AdapterRead = require('../adapter-read');
 var parsers = require('../parsers');
 var errors = require('../../errors');
