+++ lib/adapters/stdio/write.js
@@ -1,7 +1,7 @@
 'use strict';
 /* eslint-env node */
 var serializers = require('../serializers');
-var AdapterWrite = require('../../runtime/adapter-write');
+var AdapterWrite = require('../adapter-write');
 
 class WriteStdio extends AdapterWrite {
     constructor(options, params) {
