+++ lib/adapters/stochastic.js
@@ -4,7 +4,7 @@ var _ = require('underscore');
 var stoke = require('../runtime/procs/stoke');
 var values = require('../runtime/values');
 var JuttleMoment = require('../runtime/types/juttle-moment');
-var AdapterRead = require('../runtime/adapter-read');
+var AdapterRead = require('./adapter-read');
 var errors = require('../errors');
 
 class ReadStochastic extends AdapterRead {
